Rename more TextCompletion naming
