+++ dotnet/src/Connectors/Connectors.AI.HuggingFace/TextGeneration/TextGenerationResult.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.SemanticKernel.AI;
 using Microsoft.SemanticKernel.AI.TextGeneration;
 
-namespace Microsoft.SemanticKernel.Connectors.AI.HuggingFace.TextCompletion;
+namespace Microsoft.SemanticKernel.Connectors.AI.HuggingFace.TextGeneration;
 
 internal sealed class TextGenerationResult : ITextResult
 {
