+++ dotnet/src/IntegrationTests/Connectors/HuggingFace/TextGeneration/HuggingFaceTextGenerationTests.cs
@@ -5,10 +5,10 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using Microsoft.SemanticKernel.AI.TextGeneration;
-using Microsoft.SemanticKernel.Connectors.AI.HuggingFace.TextCompletion;
+using Microsoft.SemanticKernel.Connectors.AI.HuggingFace.TextGeneration;
 using Xunit;
 
-namespace SemanticKernel.IntegrationTests.Connectors.HuggingFace.TextCompletion;
+namespace SemanticKernel.IntegrationTests.Connectors.HuggingFace.TextGeneration;
 
 /// <summary>
 /// Integration tests for <see cref="HuggingFaceTextGeneration"/>.
